learnhearCore/requirements.txt
absl-py==2.0.0
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
anyascii==0.3.2
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
autobahn==23.6.2
Automat==22.10.0
av==10.0.0
Babel==2.13.1
backcall==0.2.0
bangla==0.0.2
blinker==1.7.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
boto3==1.28.3
botocore==1.31.3
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.1
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.5.2
colorama==0.4.6
coloredlogs==15.0.1
constantly==15.1.0
contextlib2==21.6.0
contourpy==1.2.0
coqpit==0.0.17
cryptography==41.0.4
ctranslate2==3.20.0
cycler==0.12.1
Cython==0.29.30
daphne==4.0.0
dateparser==1.1.8
dctorch==0.1.2
decorator==5.1.1
Django==4.2.3
django-ninja==0.22.2
django-ninja-extra==0.19.8
django-storages==1.14
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
encodec==0.1.1
executing==1.2.0
faster-whisper==0.9.0
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.44.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
funcy==2.0
future==0.18.3
g2pkk==0.1.2
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
hangul-romanize==0.1.0
huggingface-hub==0.16.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
imageio==2.32.0
incremental==22.10.0
inflect==5.6.2
injector==0.21.0
ipython==8.14.0
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonlines==1.2.0
jsonmerge==1.9.2
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
k-diffusion==0.1.1
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
librosa==0.10.0.post2
llvmlite==0.40.1
Markdown==3.5.1
MarkupSafe==2.1.2
matplotlib==3.7.3
matplotlib-inline==0.1.6
more-itertools==9.1.0
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
networkx==2.8.8
ngrok==0.8.1
nltk==3.8.1
num2words==0.5.13
numba==0.57.0
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.16.0
openai-whisper @ git+https://github.com/openai/whisper.git@0a60fcaa9b86748389a656aa013c416030287d47
packaging==23.1
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pooch==1.6.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.1.2
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.8.0
pynndescent==0.5.10
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypinyin==0.49.0
pyreadline3==3.4.1
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0
redis==5.0.1
referencing==0.31.0
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rpds-py==0.12.0
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.3
sentry-sdk==1.35.0
service-identity==23.1.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.5
sqlparse==0.4.4
stack-data==0.6.2
suno-bark==0.0.1a0
sympy==1.11.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.1.0
tifffile==2023.9.26
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.1.0
torchaudio==2.1.0
torchdiffeq==0.2.3
torchsde==0.2.6
torchvision==0.16.0
tqdm==4.64.1
trainer==0.0.31
traitlets==5.9.0
trampoline==0.1.2
transformers==4.33.3
TTS==0.20.4
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
umap-learn==0.5.4
Unidecode==1.3.6
urllib3==1.26.13
wandb==0.16.0
wcwidth==0.2.6
Werkzeug==3.0.1
yarl==1.9.2
zope.interface==6.1
